pyproject.toml
[project]
name = "css-web-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.16.1",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "gunicorn>=23.0.0",
    "httptools>=0.6.4",
    "langchain>=0.3.26",
    "langchain-cohere>=0.4.4",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langserve>=0.3.1",
    "pdfminer-six>=20250506",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.3",
    "unstructured>=0.18.5",
    "uvicorn[standard]>=0.35.0",
    "watchfiles>=1.1.0",
    "websockets>=15.0.1",
]
